VegaGraphs.jl

Create beatiful and interactive visualizations for graphs using Vega-Lite.

VegaGraphs implements graph visualization with Vega-Lite.

This library is built on top of the JuliaGraphs project. The current version supports only undirected graphs.

Package Features

Interactive graph visualization;

All layouts provided by GraphPlots;

Highly customizable visualization;

Undirected Graph Visualization

VegaGraphs.graphplot — Function

graphplot(df, layout = GraphPlot.spring_layout; kwargs...)

Input:

g is SimpleGraph of SimpleWeightedGraph.

layout is a GraphPlot layout function (e.g GraphPlot.circular_layout)

This function produces an standard graph plot. The possible arguments to be used are the following:

node_label = true

node_labelsize = 10

tooltip = true

node_size = 500

node_color = "#9ecae9"

node_sizefield = nothing

node_colorfield = nothing

node_colorscheme = "blues"

node_opacity = 1.0

edge_opacity = 0.5

width = 600

height = 400