VegaGraphs.jl - Internal Functions

This section constains the documentation concerning the internal functions used in the package. These functions are not available to the user, but might be of interest for people interested in developing new functionality to the package.

Plotting Function

These functions contain the Vega-Lite specifications for the plots.

VegaGraphs.vl_graph_plot — Function

vlgraphplot( dfnodes, dfedges, nodelabel = true, nodelabelsize = 10, tooltip = true, nodesize = 500, nodecolor = "#9ecae9", nodesizefield = nothing, nodecolorfield = nothing, nodecolorscheme = "blues", nodeopacity = 1.0, edge_opacity = 0.5, width = 600, height = 400 ) Contains the VegaLite specification for producing the plot.

Graph Function

These functions help create the DataFrames that are passed on to the Vega-Lite specifications.

VegaGraphs.get_edges — Function

getedges(g,nodex,nodey) g is a SimpleWeightedGraph, while nodex and node_y are arrays containing the position of the nodes.

Return a DataFrame containing the edge information, such as position, pairs and weights.